Audible words and settings
Added ability to use audible words. If Use Audible Words is set in
Settings then when Play is clicked we check to make sure all words
contain an audio file. When a game is played the HUD checks for this
setting to display the Audio button and we check to see if Show Words
is checked. If not then we hide the CurrentWord Text and just allow for
playing of the audio.